[install] Improve removing global git configuration.
This now clears the global git configurtation when setting the
user-specific git configuration.  This ensures that the git user name
and email are only set in one place.

This now also completely removes the user section from the global git
configuration instead of potentially leaving an empty 'user' section.

This should really be dealing with the user section more completely as
there is at least one more user configuration that should be handled
here.