[install] Create machine- and user-specific .emacs.
This creates empty machine- and user-specific .emacs files if the files
do not exist.  Some examples of information that could go into these
files are a machine-specific theme, user-specific email settings or
keys, or anything that might be specific to a particular machine's emacs
version.

It should be possible to automate modifying these files when specific
emacs behaviors are identified, though it's not obvious how that ewould
be done.