[npm] Use npm-config to set the global prefix.
This ties together the install script and the bashrc and cmdrc and makes
installing from the command line match most instructions, so the
following just works.

   npm install -g <package>

Note that cmdrc currently does not use $(npm prefix -g) as it is hard to
use the output of a command to set an environment variable.